Adding Show Properties variable
Adding a SHOW PROPERTIES boolean variable in the constants section to be fed in as arguments to anchor() and wired() as anchor_inpanel, and wired_inpanel.  The argument will be used for createNode(inpanel = anchor_inpanel) functions.

False will be hide the properties panel from appearing on anchor/wired creation

True will be show the properties panel from appearing on anchor/wired creation